+++ HoiThao_Core/Controllers/AccountsController.cs
@@ -59,7 +59,7 @@ namespace HoiThao_Core.Controllers
                 return View(account);
             }
 
-            var accountM = _accountRepository.GetById(account.Id);
+            var accountM = _accountRepository.GetById((int)(account.Id));
 
             if (string.IsNullOrEmpty(account.Password))
                 account.Password = accountM.Password;
